backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
pandas>=2.1.0
numpy>=1.26.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
scikit-learn>=1.3.0
httpx>=0.25.0
tiktoken>=0.5.0
aiofiles>=23.2.0
python-jose>=3.3.0
